requirements.txt
flask
openai==0.28
python-dotenv
SQLAlchemy
psycopg2-binary
flask-sqlalchemy
gunicorn
